the combobox preferred size should be more for classic theme fix for 118194

+++ websvc.design/src/org/netbeans/modules/websvc/design/view/ZoomManager.java
@@ -328,7 +328,7 @@ public class ZoomManager {
             // instead, give it a prototype value and then ask for the
             // preferred size, making that the maximum size
             // (make it wide enough to accomodate the '%').
-            setPrototypeDisplayValue(new Integer(10000));
+            setPrototypeDisplayValue(new Integer(100000));
             setMaximumSize(getPreferredSize());
             setEditable(true);
             Listener l = new Listener(manager);
